requirements.txt
aiogram>=3.0
google-auth>=2.0
google-auth-oauthlib>=1.0
google-api-python-client>=2.0
httpx>=0.25
python-dotenv>=1.0
